Move endpoint related functions to endpoint.rs
