+++ src/database/crud/object.rs
@@ -9,15 +9,14 @@ use diesel::{delete, insert_into, prelude::*, update};
 use r2d2::PooledConnection;
 
 use crate::api::aruna::api::storage::services::v1::{
-    AddLabelToObjectRequest, AddLabelToObjectResponse, GetObjectEndpointsRequest,
-    GetObjectEndpointsResponse, GetReferencesRequest, GetReferencesResponse, ObjectReference,
-    SetHooksOfObjectRequest, SetHooksOfObjectResponse,
+    AddLabelToObjectRequest, AddLabelToObjectResponse, GetReferencesRequest, GetReferencesResponse,
+    ObjectReference, SetHooksOfObjectRequest, SetHooksOfObjectResponse,
 };
 use crate::api::aruna::api::storage::{
     internal::v1::{Location as ProtoLocation, LocationType},
     models::v1::{
-        Endpoint as ProtoEndpoint, Hash as ProtoHash, KeyValue, Object as ProtoObject,
-        Origin as ProtoOrigin, Source as ProtoSource,
+        Hash as ProtoHash, KeyValue, Object as ProtoObject, Origin as ProtoOrigin,
+        Source as ProtoSource,
     },
     services::v1::{
         CloneObjectRequest, CloneObjectResponse, CreateObjectReferenceRequest,
@@ -64,6 +63,7 @@ pub struct ObjectDto {
     pub latest: bool,
     pub update: bool,
 }
+
 /// Implementing CRUD+ database operations for Objects
 impl Database {
     /// Creates the following records in the database to initialize a new object:
@@ -198,6 +198,7 @@ impl Database {
         })
     }
 
+    /// ToDo: Rust Doc
     pub fn finish_object_staging(
         &self,
         request: &FinishObjectStagingRequest,
@@ -577,38 +578,6 @@ impl Database {
     }
 
     ///ToDo: Rust Doc
-    pub fn get_location_endpoint(&self, location: &ObjectLocation) -> Result<Endpoint, ArunaError> {
-        let endpoint = self
-            .pg_connection
-            .get()?
-            .transaction::<Endpoint, Error, _>(|conn| {
-                let endpoint: Endpoint = endpoints
-                    .filter(database::schema::endpoints::id.eq(&location.endpoint_id))
-                    .first::<Endpoint>(conn)?;
-
-                Ok(endpoint)
-            })?;
-
-        Ok(endpoint)
-    }
-
-    ///ToDo: Rust Doc
-    pub fn get_endpoint(&self, endpoint_uuid: &uuid::Uuid) -> Result<Endpoint, ArunaError> {
-        let endpoint = self
-            .pg_connection
-            .get()?
-            .transaction::<Endpoint, Error, _>(|conn| {
-                let endpoint: Endpoint = endpoints
-                    .filter(database::schema::endpoints::id.eq(&endpoint_uuid))
-                    .first::<Endpoint>(conn)?;
-
-                Ok(endpoint)
-            })?;
-
-        Ok(endpoint)
-    }
-
-    ///ToDo: Rust Doc
     pub fn get_latest_object_revision(
         &self,
         request: GetLatestObjectRevisionRequest,
@@ -650,6 +619,7 @@ impl Database {
         Ok(all_revs)
     }
 
+    /// ToDo: Rust Docs
     pub fn get_objects(
         &self,
         request: GetObjectsRequest,
@@ -897,6 +867,7 @@ impl Database {
         // Empty response signals success
         Ok(GetReferencesResponse { references })
     }
+
     /// This clones a specific revision of an object into another collection.
     /// The cloned object is then treated like any other individual object.
     ///
@@ -1249,60 +1220,7 @@ impl Database {
         Ok(DeleteObjectResponse {})
     }
 
-    pub fn get_object_endpoints(
-        &self,
-        request: GetObjectEndpointsRequest,
-    ) -> Result<GetObjectEndpointsResponse, ArunaError> {
-        let parsed_object_id = uuid::Uuid::parse_str(&request.object_id)?;
-        // Transaction time
-        let obj_eps = self
-            .pg_connection
-            .get()?
-            .transaction::<(Vec<Endpoint>, uuid::Uuid), Error, _>(|conn| {
-                // Get collection_object association of original object
-                let locations = object_locations
-                    .filter(database::schema::object_locations::object_id.eq(&parsed_object_id))
-                    .load::<ObjectLocation>(conn)?;
-                let endpoint_ids = locations.iter().map(|e| e.endpoint_id).collect::<Vec<_>>();
-
-                let default_id = locations.iter().find(|e| e.is_primary).unwrap();
-
-                Ok((
-                    endpoints
-                        .filter(database::schema::endpoints::id.eq_any(&endpoint_ids))
-                        .load::<Endpoint>(conn)?,
-                    default_id.endpoint_id,
-                ))
-            })?;
-
-        let mapped_ep = obj_eps
-            .0
-            .iter()
-            .map(|ep| {
-                let is_default = ep.id == obj_eps.1;
-
-                ProtoEndpoint {
-                    id: ep.id.to_string(),
-                    ep_type: ep.endpoint_type as i32,
-                    name: ep.name.to_string(),
-                    proxy_hostname: ep.proxy_hostname.to_string(),
-                    internal_hostname: ep.internal_hostname.to_string(),
-                    documentation_path: ep
-                        .documentation_path
-                        .as_ref()
-                        .map(|e| e.to_string())
-                        .unwrap_or_default(),
-                    is_public: ep.is_public,
-                    is_default,
-                }
-            })
-            .collect::<Vec<_>>();
-
-        Ok(GetObjectEndpointsResponse {
-            endpoints: mapped_ep,
-        })
-    }
-
+    /// ToDo: Rust Doc
     pub fn add_label_to_object(
         &self,
         request: AddLabelToObjectRequest,
@@ -1331,6 +1249,8 @@ impl Database {
             object: Some(updated_objects.try_into()?),
         })
     }
+
+    /// ToDo: Rust Doc
     pub fn set_hooks_of_object(
         &self,
         request: SetHooksOfObjectRequest,
@@ -1542,6 +1462,7 @@ pub fn get_all_revisions(
 
     Ok(all_revision_objects)
 }
+
 /// Query all references for a specific object optionally with revisions
 ///
 /// ## Arguments:
@@ -1606,6 +1527,7 @@ pub fn get_all_references(
             .load::<CollectionObject>(conn)
     }
 }
+
 /// Helper function to query the current database version of an object "ObjectDto"
 /// ObjectDto can be transformed to gRPC objects
 ///
